clean and fuse normalization layers tests
